+++ packages/base-docs/package.json
@@ -35,6 +35,7 @@
         "esbuild": "node ./esbuild.config.mjs && tsc -p tsconfig.esbuild.json"
     },
     "dependencies": {
+        "@ctrl/tinycolor": "^4.0.2",
         "@univerjs/base-render": "workspace:*",
         "@univerjs/base-sheets": "workspace:*",
         "@univerjs/base-ui": "workspace:*",
