+++ packages/base-docs/src/commands/commands/__tests__/clipboard.command.spec.ts
@@ -6,9 +6,16 @@ import {
     TextSelectionManagerService,
 } from '@univerjs/base-docs';
 import {
+    IInnerCutCommandParams,
+    IInnerPasteCommandParams,
+    InnerCutCommand,
+    InnerPasteCommand,
+} from '@univerjs/base-docs/commands/commands/clipboard.inner.command.js';
+import {
     BooleanNumber,
     ICommand,
     ICommandService,
+    IDocumentData,
     IStyleBase,
     IUniverInstanceService,
     UndoCommand,
@@ -17,13 +24,62 @@ import {
 import { Injector } from '@wendellhu/redi';
 import { afterEach, beforeEach, describe, expect, it } from 'vitest';
 
-import {
-    IInnerCutCommandParams,
-    IInnerPasteCommandParams,
-    InnerCutCommand,
-    InnerPasteCommand,
-} from '../clipboard.command';
 import { createCommandTestBed } from './create-command-test-bed';
+
+const TEST_DOCUMENT_DATA_EN: IDocumentData = {
+    id: 'test-doc',
+    body: {
+        dataStream: 'What’s New in the 2022\r Gartner Hype Cycle for Emerging Technologies\r\n',
+        textRuns: [
+            {
+                st: 0,
+                ed: 22,
+                ts: {
+                    bl: BooleanNumber.FALSE,
+                    fs: 24,
+                    cl: {
+                        rgb: 'rgb(0, 40, 86)',
+                    },
+                },
+            },
+            {
+                st: 23,
+                ed: 68,
+                ts: {
+                    bl: BooleanNumber.TRUE,
+                    fs: 24,
+                    cl: {
+                        rgb: 'rgb(0, 40, 86)',
+                    },
+                },
+            },
+        ],
+        paragraphs: [
+            {
+                startIndex: 22,
+            },
+            {
+                startIndex: 68,
+                paragraphStyle: {
+                    spaceAbove: 20,
+                    indentFirstLine: 20,
+                },
+            },
+        ],
+        sectionBreaks: [],
+        customBlocks: [],
+    },
+    documentStyle: {
+        pageSize: {
+            width: 594.3,
+            height: 840.51,
+        },
+        marginTop: 72,
+        marginBottom: 72,
+        marginRight: 90,
+        marginLeft: 90,
+    },
+};
 
 describe('test cases in clipboard', () => {
     let univer: Univer;
@@ -55,7 +111,7 @@ describe('test cases in clipboard', () => {
     }
 
     beforeEach(() => {
-        const testBed = createCommandTestBed();
+        const testBed = createCommandTestBed(TEST_DOCUMENT_DATA_EN);
         univer = testBed.univer;
         get = testBed.get;
